功能 #2283 试卷应该具有３种状态，draft/published/closed。

+++ src/Topxia/WebBundle/Controller/QuizQuestionTestController.php
@@ -91,6 +91,12 @@ class QuizQuestionTestController extends BaseController
 
 		$testPaper = $request->query->all();
 
+		$paper = $this->getTestService()->getTestPaper($testPaper['testPaperId']);
+
+		if ($paper['status'] != 'draft'){
+			throw $this->createAccessDeniedException('只有草稿状态的试卷可以修改');
+		}
+
 	    if ($request->getMethod() == 'POST') {
 
 	    	$testPaper = $request->request->all();
@@ -124,6 +130,12 @@ class QuizQuestionTestController extends BaseController
 
 		if(empty($testPaper['testPaperId'])){
 			throw $this->createNotFoundException('缺少参数');
+		}
+
+		$paper = $this->getTestService()->getTestPaper($testPaper['testPaperId']);
+
+		if ($paper['status'] != 'draft'){
+			throw $this->createAccessDeniedException('只有草稿状态的试卷可以修改');
 		}
 
 		$this->getTestService()->deleteItemsByTestPaperId($testPaper['testPaperId']);
@@ -161,6 +173,10 @@ class QuizQuestionTestController extends BaseController
 		$testPaper = $this->getTestService()->getTestPaper($testPaperId);
 		if(empty($testPaper)){
 			throw $this->createNotFoundException('试卷不存在');
+		}
+
+		if ($testPaper['status'] != 'draft'){
+			throw $this->createAccessDeniedException('只有草稿状态的试卷可以修改');
 		}
 
 		$flag = $request->query->get('flag');
@@ -347,6 +363,10 @@ class QuizQuestionTestController extends BaseController
 
         if (empty($testPaper)) {
             throw $this->createNotFoundException();
+        }
+
+        if ($testPaper['status'] != 'draft'){
+        	throw $this->createAccessDeniedException('只有草稿状态的试卷可以删除');
         }
 
         $this->getTestService()->deleteTestPaper($testPaperId);
